changes to nm.tex
1) million not Million
2) removed the spacing symbold from  \cursedforest\  -- the definition ends in an \xspace which will space it properly
3) fixed double quotes so LaTeX  will draw left and right quotation marks
4) make n and p italic
5) we have inconsistent capitalization on Bone Mineral Density, Bone mineral density, bone mineral density
I think we get rid of all the capitals
6) datasets or data sets ?
7)  We have a number of instances where we start a sentence with a reference. This is OK in an author-year style but with a numeric style I have had editors complain about it. Do we want to change this? i.e.
[17] demonstrates ...

8) need more information about the mixed model, supplementary?

9)
We limit the number of executors
to 12 that are available to CursedForest to keep the comparison fair for all methods.

should this be

We limit the number of executors that are available to CursedForest
to 12 to keep the comparison fair for all methods.

or does it mean something else?